Move normalizeToUnitCube to transformNode instead of abstract mesh and add predicate to exclude sub objects when scaling
